requirements.txt
click==7.1.1
Flask==1.1.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pip==20.0.2
setuptools==46.0.0
Werkzeug==1.0.0
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
idna==2.9
importlib-metadata==1.5.0
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
requests==2.23.0
six==1.14.0
todoist-python==8.1.1
urllib3==1.25.8
wcwidth==0.1.8
zipp==3.1.0